clean up jupyter tests
