+++ src/test/jupyter/jupyter.codeHelper.test.ts
@@ -1,8 +1,3 @@
-//
-// Note: This example test is leveraging the Mocha test framework.
-// Please refer to their documentation on https://mochajs.org/ for help.
-//
-// The module 'assert' provides assertion methods from node
 import * as assert from 'assert';
 import * as vscode from 'vscode';
 import * as path from 'path';
