+++ src/test/jupyter/jupyterKernel.test.ts
@@ -1,64 +1,11 @@
-//
-// Note: This example test is leveraging the Mocha test framework.
-// Please refer to their documentation on https://mochajs.org/ for help.
-//
-// The module 'assert' provides assertion methods from node
 import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
-import { initialize, TEST_TIMEOUT } from './../initialize';
+import { MockOutputChannel } from './mocks';
+import { initialize } from './../initialize';
 import { JupyterClientAdapter } from '../../client/jupyter/jupyter_client/main';
 import { KernelShutdownError } from '../../client/jupyter/common/errors';
 import { createDeferred } from '../../client/common/helpers';
 import { JupyterClientKernel } from '../../client/jupyter/jupyter_client-Kernel';
 import { KernelspecMetadata } from '../../client/jupyter/contracts';
-import * as vscode from 'vscode';
-
-export class MockOutputChannel implements vscode.OutputChannel {
-    constructor(name: string) {
-        this.name = name;
-        this.output = '';
-        this.timeOut = setTimeout(() => {
-            console.log(this.output);
-            this.writeToConsole = true;
-            this.timeOut = null;
-        }, TEST_TIMEOUT - 1000);
-    }
-    private timeOut: number;
-    name: string;
-    output: string;
-    isShown: boolean;
-    private writeToConsole: boolean;
-    append(value: string) {
-        this.output += value;
-        if (this.writeToConsole) {
-            console.log(value);
-        }
-    }
-    appendLine(value: string) {
-        this.append(value); this.append('\n');
-        if (this.writeToConsole) {
-            console.log(value);
-            console.log('\n');
-        }
-    }
-    clear() { }
-    show(preservceFocus?: boolean): void;
-    show(column?: vscode.ViewColumn, preserveFocus?: boolean): void;
-    show(x?: any, y?: any): void {
-        this.isShown = true;
-    }
-    hide() {
-        this.isShown = false;
-    }
-    dispose() {
-        if (this.timeOut) {
-            clearTimeout(this.timeOut);
-            this.timeOut = null;
-        }
-    }
-}
 
 suite('Jupyter Kernel', () => {
     suiteSetup(() => initialize());
